Is the commit txt the same or did it not commit the first time ?
I have now created /groups
It displays all the groups (city meetup groups)
I display the different groups using the blog post as template. This is a hack and must be fixed.
It is now possible to click on a group and see the homepage (grouphome.js/jade)for that group eg: /group/bergen
The group homepage is a copy of the index.jade/index.js and display the last meetup for the group.
The bootom part of the page now displays fields from the group (their slogan, description, twitter, and so on)

what is missing is to complete the fields and make it look good.
the group homepage has its own top image and bottom image. I do not know how to do the css/html to do this.
The fields are there. So its a matter of putting them in css/html